bug(#349): read the changelog for release notes, drop the master push
Rultor already creates the GitHub Release and master is protected, so
release.yml's gh release create hit a duplicate and its changelog push
to master could not land. Read the released version's changelog section
and set the notes with gh release edit (falling back to create); promote
the changelog in the pre-tag PR instead of from CI. Rename the script to
changelog-notes.js (read-only extractor). Promote 0.0.8, which shipped,
and prepare 0.0.9.
