+++ .github/workflows/release.yml
@@ -17,9 +17,6 @@ jobs:
     runs-on: ubuntu-24.04
     steps:
       - uses: actions/checkout@v4
-        with:
-          ref: master
-          fetch-depth: 0
       - uses: actions/setup-node@v4
         with:
           node-version: 22
@@ -39,23 +36,12 @@ jobs:
         run: |
           npm install -g npm@latest
           npm publish --provenance --access public
-      - name: Promote the changelog and collect the release notes
-        env:
-          VERSION: ${{ github.ref_name }}
-        run: |
-          today="$(date +%F)"
-          node scripts/changelog-release.js "${VERSION}" "${today}" > notes.md
-      - name: Create the GitHub release
+      - name: Set the GitHub release notes from the changelog
         env:
           GH_TOKEN: ${{ github.token }}
           VERSION: ${{ github.ref_name }}
-        run: gh release create "${VERSION}" --title "${VERSION}" --notes-file notes.md
-      - name: Commit the promoted changelog to master
-        env:
-          VERSION: ${{ github.ref_name }}
         run: |
-          git config user.name 'github-actions[bot]'
-          git config user.email '41898282+github-actions[bot]@users.noreply.github.com'
-          git add CHANGELOG.md
-          git commit -m "changelog: release ${VERSION}"
-          git push origin HEAD:master
+          node scripts/changelog-notes.js "${VERSION}" > notes.md
+          gh release edit "${VERSION}" --notes-file notes.md \
+            || gh release create "${VERSION}" --title "${VERSION}" \
+              --notes-file notes.md
